chore(city): added back button to the city component
