+++ src/styles/city.scss
@@ -60,6 +60,10 @@
   margin-right: auto;
 }
 
+.recently-visited {
+  border: 2px solid $color-brand-2;
+}
+
 .city-date {
   font-size: 1rem;
 }
